Added in tests and scenario from Tanya
Added in three element tests than Tanya and I worked on previously, and also added in scenario 25 that Tanya sent me. IUArrayList has been updated with the iterator. There are a few issues that have occured with the arraylist that didn't before. I am working on this and will hopefully have it fixed soon.